Update extra pages type

+++ src/to_my_algo_transaction.rs
@@ -337,7 +337,7 @@ struct MyAlgoApplicationCreateTransactionFields {
     appGlobalInts: u64,
     appGlobalByteSlices: u64,
     appOnComplete: u32,
-    extraPages: u64,
+    extraPages: u32,
 }
 
 #[skip_serializing_none]
